refactor: extract signing-path into standalone feature
Move path-model, graph-model, path-validation and the signing-path UI
out of features/drafts into features/signing-path so the picker can be
reused (next consumer: edit flexible multisig). Drafts now imports the
public surface via @/features/signing-path. PathNode stays in
domains/backend (Draft API contract). Auto-reset on chain change is left
in create-draft-model — it's a drafts policy, not a signing-path concern.
StepIndicator was misfiled under signing-path/; relocated to
drafts/components/.
