+++ src/renderer/features/signing-path/lib/path-validation.test.ts
