+++ src/renderer/features/signing-path/lib/path-validation.ts
