+++ src/renderer/features/signing-path/model/graph-model.ts
