+++ src/renderer/features/signing-path/model/path-model.test.ts
