+++ src/renderer/features/signing-path/ui/EllipsisCard.tsx
