+++ src/renderer/features/signing-path/ui/NextOptionRow.test.tsx
