+++ src/renderer/features/signing-path/ui/PathArrow.tsx
