+++ src/renderer/features/signing-path/ui/PathCard.tsx
