+++ src/renderer/features/signing-path/ui/SectionCard.tsx
