+++ src/renderer/features/signing-path/ui/path-views.ts
